+++ components/footer.ejs
@@ -1,6 +1,8 @@
-<div class="fixed-bottom">
-  <div class="text-center">
-    2020 Duckey.Kim :<a href="mailto:">duckey.kim@gmail.com</a>
+<div class="text-monospace">
+  Copyright(c) 2020 Duckey.Kim all rights reserved.
+  <div>
+    <a href="mailto:#">duckey-kdh@google.com</a>
+    <a href="https://github.com/duckey-kim"
+      style="text-decoration: none; color: brown;">https://github.com/duckey-kim</a>
   </div>
-
 </div>
