+++ components/head.ejs
@@ -3,15 +3,13 @@
 
 <link href="http://fonts.googleapis.com/css?family=Open+Sans:400italic,600italic,400,600,300" rel="stylesheet"
   type="text/css" />
-<link rel="stylesheet" type="text/css"
-  href="https://cdn.rawgit.com/innks/NanumSquareRound/master/nanumsquareround.min.css">
 
 <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css"
   integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh" crossorigin="anonymous" />
-
+<link rel="stylesheet" href="/public/css/style.css" />
 <script src="https://cdn.firebase.com/libs/firebaseui/3.5.2/firebaseui.js"></script>
 <link type="text/css" rel="stylesheet" href="https://cdn.firebase.com/libs/firebaseui/3.5.2/firebaseui.css" />
-<link rel="stylesheet" href="/public/css/style.css" />
+
 <script>
   initApp = function () {
     setCookie("userName", "");
@@ -44,12 +42,8 @@
 
   function signOut() {
     alert("logout!!");
-    firebase.auth().signOut().then(function () {
-      console.log("sign out!!");
-      window.location.reload();
-    }).catch(function (error) {
-      console.log(error);
-    });
+    firebase.auth().signOut();
+    window.location.reload();
   }
 </script>
 <!-- ckeditor -->
