+++ components/jumbotron.ejs
@@ -1,10 +1,9 @@
-<div class="jumbotron" style="
+<div class="jumbotron mb-2" style="
         background-image: url('/public/img/mountains.png');
         background-repeat: no-repeat;
         background-size: cover;
         background-position: center;
-        margin-top: 3em;
-        margin-bottom:1em;
-        min-height: 200px;
+        margin-top: 50px;
+        min-height: 100px;
       ">
 </div>
