+++ views/account.ejs
@@ -1,10 +1,5 @@
 <div id="firebaseui-auth-container"></div>
-<div class="d-flex justify-content-center">
-  <div class="spinner-border" id="loader" role="status">
-    <span class="sr-only">Loading...</span>
-  </div>
-</div>
-
+<div id="loader">LOADING</div>
 
 <%- contentFor('account') %>
