+++ views/boards/category.ejs
@@ -1,29 +1,54 @@
-<div class="row">
-  <div class="col-12">
-    <div class="categoryBox mt-1">
-      <div class="row bg-pink borded rounded category">
-        <%= category.toUpperCase() %>
-      </div>
+<div class="table-responsive-lg" style="font-size:14px">
+  <table class="table table-hover table-borderless  text-left">
+    <colgroup>
+      <col width='10%' />
+      <col width='*%' />
+      <col width='20%' />
+      <col width='20%' />
+    </colgroup>
+    <thead>
+      <tr>
+        <th colspan="4" id="boardName" class="bg-info"><%= category %></th>
+      </tr>
+      <tr>
+        <th scope="column" class="postTitle">글번호</th>
+        <th scope="column" class=" postUser">제목</th>
+        <th scope="column" class="postLike">작성자</th>
+        <th scope="column" class="postDate">일자</th>
+      </tr>
+    </thead>
+    <tbody>
+    <tbody>
       <% for(var i=0; i<data.length; i++) { %>
-      <div class="row mt-1" id="boards" onclick="location.href='/boards/<%=category%>/posts/<%= data[i].postnum%>'">
-        <div class="col-12 col-md-8  bold text-left text-truncate">
-          <%=data[i].posttitle%>
-        </div>
-        <div class="col-6 col-md-2 fontSizeDown">
-          <%=data[i].postuser%>
-        </div>
-        <div class="col-6 col-md-2 fontSizeDown  text-right text-truncate">
-          <%=data[i].uploadtime%>
-        </div>
-      </div>
-      <% }%>
-      <% if(userName) { %>
-      <div class="row">
-        <div type=" button" id="writeButton" onclick="location.href='/boards/<%= category %>/posts/new'"
-          class="col-12 offset-md-10 col-md-2 btn bg-beige btn-md mt-2">
-          Write</div>
-      </div>
-      <% } %>
-    </div>
-  </div>
+      <tr onclick="location.href='/boards/<%=category%>/posts/<%= data[i].postnum%>'">
+        <td scope="row"><%=data.length-i%></a></td>
+        <td><%=data[i].posttitle%></a></td>
+        <td><%=data[i].postuser%></td>
+
+        <td><%= data[i].uploadtime%></td>
+      </tr>
+      <%
+        }
+       %>
+    </tbody>
+  </table>
+
 </div>
+<% if(userName) { %>
+<div type=" button" id="writeButton" onclick="location.href='/boards/<%= category %>/posts/?action=new_post'"
+  class="btn btn-primary btn-sm float-right">
+  Write</div>
+<% } %>
+
+
+<!-- board container end -->
+
+<!-- pagination set -->
+
+<ul class="pagination justify-content-center">
+  <li class="page-item"><a class="page-link" href="#">Previous</a></li>
+  <li class="page-item"><a class="page-link" href="#">1</a></li>
+  <li class="page-item"><a class="page-link" href="#">2</a></li>
+  <li class="page-item"><a class="page-link" href="#">3</a></li>
+  <li class="page-item"><a class="page-link" href="#">Next</a></li>
+</ul>
