+++ views/boards/index.ejs
@@ -1,25 +1,34 @@
-<div class="row">
-  <% data.forEach((posts,category)=>{%>
-  <div class="col-12 col-md-6">
-    <div class="indexBox mt-1">
-      <div class="row borded rounded bg-pink indexCategory" onclick="location.href='/boards/<%= category %>/posts'">
-        <%= category.toUpperCase()%>
-      </div>
-      <div class="row">
-        <div class="col-12">
+<div class="table-responsive-xl">
+  <div class="row">
+    <% data.forEach((posts,category)=>{%>
+    <div class="col-md-6">
+      <table class="table table-borderless table-hover">
+        <colgroup>
+          <col width='*%' />
+          <col width='33%' />
+          <col width='33%' />
+        </colgroup>
+        <thead>
+          <tr>
+            <th colspan="3" class="bg-info" onclick="location.href='/boards/<%= category %>'">
+              <%= category%>
+            </th>
+          </tr>
+        </thead>
+        <tbody>
           <% posts.forEach((post)=>{ %>
-          <div class="row mt-1" id="boards" onclick="location.href='/boards/<%=category%>/posts/<%= post.postnum%>'">
-            <div class="col-8 text-truncate">
-              <%=post.posttitle%>
-            </div>
-            <div class="col-4 text-right text-truncate">
-              <%=post.postuser%>
-            </div>
-          </div>
-          <% }) %>
-        </div>
-      </div>
+          <tr onclick="location.href='/boards/<%=category%>/posts/<%= post.postnum%>'">
+
+            <td scope="row"><%=post.posttitle%></a></td>
+            <td><%=post.postuser%></td>
+            <td><%=(post.uploadtime.toString())%></td>
+          </tr>
+          <%
+            })
+           %>
+        </tbody>
+      </table>
     </div>
+    <% })%>
   </div>
-  <% })%>
 </div>
